Split authentication out into its own file
